+++ CHANGELOG.md
@@ -16,8 +16,11 @@ New methods:
 
 - Added ``minimum_clearance`` method from shapely to GeoSeries/GeoDataframe (#2989).
 
+API changes:
+- Added support for ``mask`` keyword for pyogrio engine for pyogrio >= 0.7.0
+
 Bug fixes:
-- Fix `GeoDataFrame.merge()` incorrectly returning a `DataFrame` instead of a 
+- Fix `GeoDataFrame.merge()` incorrectly returning a `DataFrame` instead of a
   `GeoDataFrame` when the `suffixes` argument is applied to the active
   geometry column (#2933).